EBMS/requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.4.4
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
sqlparse==0.4.4
tzdata==2024.1